# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "safehaven",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/map/comments": {
      "post": {
        "tags": [
          "map"
        ],
        "operationId": "viewer_new_comment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Comment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicComment"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/map/entities": {
      "post": {
        "tags": [
          "map"
        ],
        "operationId": "viewer_new_entity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicNewEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicNewEntityResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/map/entities/{id}": {
      "post": {
        "tags": [
          "map"
        ],
        "operationId": "viewer_fetch_entity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchEntityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchedEntity"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/map/search": {
      "post": {
        "tags": [
          "map"
        ],
        "operationId": "viewer_search_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of entities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewerCachedEntitiesWithPagination"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/map/view": {
      "post": {
        "tags": [
          "map"
        ],
        "operationId": "viewer_view_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "List of entities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitiesAndClusters"
                }
              }
            }
          },
          "401": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EntitiesAndClusters": {
        "type": "object",
        "required": [
          "entities",
          "clusters"
        ],
        "properties": {
          "clusters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cluster"
            }
          },
          "entities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ViewerCachedEntity"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error_code"
        ],
        "properties": {
          "details": {
            "type": "string",
            "nullable": true
          },
          "error_code": {
            "type": "string"
          }
        }
      },
...
      "FetchEntityRequest": {
        "type": "object",
        "required": [
          "active_categories",
          "active_required_tags",
          "active_hidden_tags",
          "enums_constraints"
        ],
        "properties": {
          "active_categories": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "active_hidden_tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "active_required_tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "enums_constraints": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {}
            }
          }
        }
      },
      "FetchedEntity": {
        "type": "object",
        "required": [
          "entity",
          "comments",
          "parents",
          "children"
        ],
        "properties": {
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicListedEntity"
            }
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicComment"
            }
          },
          "entity": {
            "$ref": "#/components/schemas/PublicEntity"
          },
          "parents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicListedEntity"
            }
          }
        }
      },
...
      "NewCommentRequest": {
        "type": "object",
        "required": [
          "comment"
        ],
        "properties": {
          "comment": {
            "$ref": "#/components/schemas/PublicNewComment"
          },
          "hcaptcha_token": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "PublicComment": {
        "type": "object",
        "required": [
          "id",
          "author",
          "text",
          "data",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "author": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "data": {},
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "text": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PublicNewEntityRequest": {
        "type": "object",
        "required": [
          "entity"
        ],
        "properties": {
          "comment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicNewComment"
              }
            ],
            "nullable": true
          },
          "entity": {
            "$ref": "#/components/schemas/PublicNewEntity"
          },
          "hcaptcha_token": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PublicNewEntityResponse": {
        "type": "object",
        "required": [
          "entity"
        ],
        "properties": {
          "comment": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicComment"
              }
            ],
            "nullable": true
          },
          "entity": {
            "$ref": "#/components/schemas/PublicEntity"
          }
        }
      },
...
      "SearchRequest": {
        "type": "object",
        "required": [
          "search_query",
          "family_id",
          "page",
          "page_size",
          "active_categories",
          "active_required_tags",
          "active_hidden_tags",
          "require_locations",
          "enums_constraints"
        ],
        "properties": {
          "active_categories": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "active_hidden_tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "active_required_tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "enums_constraints": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {}
            }
          },
          "family_id": {
            "type": "string",
            "format": "uuid"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "page_size": {
            "type": "integer",
            "format": "int64"
          },
          "require_locations": {
            "type": "boolean"
          },
          "search_query": {
            "type": "string"
          }
        }
      },
...
      "ViewRequest": {
        "type": "object",
        "required": [
          "family_id",
          "xmin",
          "ymin",
          "xmax",
          "ymax",
          "zoom_level",
          "active_categories",
          "active_required_tags",
          "active_hidden_tags",
          "enums_constraints"
        ],
        "properties": {
          "active_categories": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "active_hidden_tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "active_required_tags": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "enums_constraints": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {}
            }
          },
          "family_id": {
            "type": "string",
            "format": "uuid"
          },
          "xmax": {
            "type": "number",
            "format": "double"
          },
          "xmin": {
            "type": "number",
            "format": "double"
          },
          "ymax": {
            "type": "number",
            "format": "double"
          },
          "ymin": {
            "type": "number",
            "format": "double"
          },
          "zoom_level": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "ViewerCachedEntitiesWithPagination": {
        "type": "object",
        "required": [
          "entities",
          "total_results",
          "total_pages",
          "response_current_page"
        ],
        "properties": {
          "entities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ViewerSearchedCachedEntity"
            }
          },
          "response_current_page": {
            "type": "integer",
            "format": "int64"
          },
          "total_pages": {
            "type": "integer",
            "format": "int64"
          },
          "total_results": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  }
}